updates itinerary into Stack and Screen, creates ItineraryPage to do so, and updates to use CityDropdown so it is the same as the other pages
